lab 2 initalizing project structure.
