+++ feature-pipeline.yml
@@ -1,0 +1,22 @@
+trigger:
+  - feature/*
+
+pool:
+  vmImage: 'ubuntu-latest'
+
+steps:
+
+  - template: templates/install-dependencies.yaml
+
+  - script: |
+      robot --include new-feature --output results/logs/output_feature_test.xml --log results/logs/log_feature_test.html --report results/logs/report_feature_test.html tests/
+    displayName: 'Run Tests tagged with new-feature'
+
+  # Publish the Robot Framework reports as pipeline artifacts.
+  - task: PublishPipelineArtifact@1
+    inputs:
+      targetPath: '$(System.DefaultWorkingDirectory)/results/logs/'
+      artifactName: 'robot-reports-feature'
+    displayName: 'Publish Robot Framework Reports'
+
+
