+++ templates/install-dependencies.yaml
@@ -1,0 +1,4 @@
+steps:
+  - script: |
+      pip install --upgrade robotframework-seleniumlibrary
+    displayName: 'Install Robot Framework and dependencies'
